+++ packages/lens-react/lib/components/Datatable/components/DatatableList.tsx
@@ -49,6 +49,7 @@ const DatatableList = () => {
     localContext,
     allowedScopes,
     variantClasses,
+    dataCallback,
   } = useDatatableContext();
 
   const { flatData, fetchNextPage, isFetching, hasNextPage, refetch } =
@@ -64,6 +65,7 @@ const DatatableList = () => {
           allowedScopes.length > 0 && { scopes: allowedScopes }),
       },
       globalFilter: searchQuery,
+      dataCallback: dataCallback,
     });
 
   useEffect(() => {