+++ packages/lens-react/lib/components/Datatable/hooks/getColumnPinningStyles.ts
@@ -2,11 +2,10 @@ import { CSSProperties } from "react";
 import type { Column } from "@tanstack/react-table";
 
 const getColumnPinningStyles = (
-  column: Column<any>,
-  isHeader?: boolean
+  column: Column<any>
+  // isHeader?: boolean
 ): CSSProperties => {
   const isPinned = column.getIsPinned();
-  isHeader; // Remove this line later
   // const isLastLeftPinnedColumn =
   //   isPinned === "left" && column.getIsLastColumn("left");
   // const isFirstRightPinnedColumn =